UART_Tx
Run our verification and corrected all the mistakes in the Tx UART Protocol  code.

It's not much we just send to letters , "u" and "R" and they are sent correctly and with the debouncing on.

Also increased the baud rate , cause we could ! Also better cause it makes the verification time smaller therefore the ammount of data produced in each scenario.

We are going to create a prototype  in the Nexys 4 Artix-7 FPGA where we will send data via a COM  port to our PUTTY cmd with a 115200 baud rate using the switches to and a button to send 10-bit packets of data   .