FPGA Prototyping
Today we completed a small prototype in which we send a message using 8 switches  to simulate 8-bit input  and a button to simulate the send_enable.

To send the message after the bitstream installanion on the board we set up a COM on the specific COM that our board is presented in the Device Manager. We set the baud rate at 115200 bps and we are ready to send data !